Add reports to ingest
