+++ state_fin_ingest/mi/README.md
@@ -12,7 +12,7 @@ The following contribution files required are:
 * 2020_mi_cfr_contributions_01.zip
 * 2020_mi_cfr_contributions_02.zip
 
-Candidate Summary ZIP file is available here: https://miboecfr.nictusa.com/cfr/dumpall/micfrcansummary.txt.zip
+Candidate Report Summary ZIP file is available here: https://miboecfr.nictusa.com/cfr/dumpall/micfrcansummary.txt.zip
 
 Place all 6 of those ZIP files in the `input/mi` directory (create the `mi` directory if missing)
 
@@ -22,3 +22,5 @@ We build a list of candidate committees from the report data (only looking at ca
 There is some weirdness in the tab-delimited files:
 1. The header only appears in the files ending in _00.txt
 2. In one of the files, there's a record that somehow has an extra field partially duplicated causing the record to have an unexpected amount of columns. We manually skip this record
+
+In addition to loading contributions, we also load report data provide by that summary document
